Fixes for DataObject.QueryGetData
Callers are allowed to call QueryGetData with a dwAspect of 0; they use this to mean "don't care" (e.g. Windows Explorer will do this).

Treat anyone who calls with zero (or negative) dwAspect as "don't care".  To that end: store dwAspect in local variable, and patch any negative value to 0.

Other bugfix: A caller is allowed to pass an lIndex of -1, also meaning "don't care". Originally we would use the same rule: if lIndex is negative, then take it to mean "don't care". But even more than that, the shell itself will advertise multiple FileContents with the same lIndex of 0.

The reality is that callers are never supposed to call QueryGetData with any index other than -1 (meaning don't care). MSDN reminds us of all this in the current documentation of QueryGetData: there is no point is passing anything other than -1 *"(returning error `DV_E_LINDEX`
- Invalid value for lindex; currently, only -1 is supported.)*. This means that it is wrong to check lIndex at all.

We could be rigid and inflexible to poorly written applications, and also return DV_E_LINDEX if they called with anything besides -1, but this needlessly breaks other applications.