Fix OnChange event with post nil if the note is the same
I’m submitting a fix for the OnChange event, which incorrectly set the node to nil when clicking on an already selected node.
This behavior is incorrect. The event should always provide the node being switched to; otherwise, nil would mean that no node is selected at all. Moreover, this worked differently in earlier versions of VT, where the parameter was never nil.